Create a real UMD script
