+++ lib/amd/lib/countries/andorra.js
@@ -6,8 +6,8 @@ define(["require", "exports"], function (require, exports) {
         name: 'Andorra',
         codes: ['AD', 'AND', '020'],
         calcFn: (vat) => {
-            let firstLetter = vat.slice(0, 1).toUpperCase();
-            let number = parseInt(vat.slice(1, 7));
+            const firstLetter = vat.slice(0, 1).toUpperCase();
+            const number = parseInt(vat.slice(1, 7), 10);
             if (vat.length !== 8) {
                 return false;
             }
