+++ lib/commonjs/lib/countries/andorra.js
@@ -6,7 +6,7 @@ exports.andorra = {
     codes: ['AD', 'AND', '020'],
     calcFn: function (vat) {
         var firstLetter = vat.slice(0, 1).toUpperCase();
-        var number = parseInt(vat.slice(1, 7));
+        var number = parseInt(vat.slice(1, 7), 10);
         if (vat.length !== 8) {
             return false;
         }
