--- lib/umd/lib/countries/belgium.js
@@ -1,27 +1,0 @@
-(function (factory) {
-    if (typeof module === "object" && typeof module.exports === "object") {
-        var v = factory(require, exports);
-        if (v !== undefined) module.exports = v;
-    }
-    else if (typeof define === "function" && define.amd) {
-        define(["require", "exports"], factory);
-    }
-})(function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    exports.belgium = void 0;
-    exports.belgium = {
-        name: 'Belgium',
-        codes: ['BE', 'BEL', '056'],
-        calcFn: (_vat) => {
-            const vat = String(_vat).replace('BE', '');
-            const check = 97 - (Number(vat.slice(0, 8)) % 97);
-            return check === Number(vat.slice(8, 10));
-        },
-        rules: {
-            multipliers: {},
-            regex: [/^(BE){1}([0-1]{1}\d{9})$/]
-            // regex: [/^(BE){0,1}([0-1]{1}\d{9})$/] // BE: Optional Country Code
-        }
-    };
-});
