--- lib/umd/lib/countries/denmark.js
@@ -1,30 +1,0 @@
-(function (factory) {
-    if (typeof module === "object" && typeof module.exports === "object") {
-        var v = factory(require, exports);
-        if (v !== undefined) module.exports = v;
-    }
-    else if (typeof define === "function" && define.amd) {
-        define(["require", "exports"], factory);
-    }
-})(function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    exports.denmark = void 0;
-    exports.denmark = {
-        name: 'Denmark',
-        codes: ['DK', 'DNK', '208'],
-        calcFn: (vat) => {
-            let total = 0;
-            for (let i = 0; i < 8; i++) {
-                total += Number(vat.charAt(i)) * exports.denmark.rules.multipliers.common[i];
-            }
-            return total % 11 === 0;
-        },
-        rules: {
-            multipliers: {
-                common: [2, 7, 6, 5, 4, 3, 2, 1]
-            },
-            regex: [/^(DK)(\d{8})$/]
-        }
-    };
-});
