--- lib/umd/lib/countries/europe.js
@@ -1,26 +1,0 @@
-(function (factory) {
-    if (typeof module === "object" && typeof module.exports === "object") {
-        var v = factory(require, exports);
-        if (v !== undefined) module.exports = v;
-    }
-    else if (typeof define === "function" && define.amd) {
-        define(["require", "exports"], factory);
-    }
-})(function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    exports.europe = void 0;
-    exports.europe = {
-        name: 'Europe',
-        codes: ['EU', 'EUR', '000'],
-        calcFn: () => {
-            // We know little about EU numbers apart from the fact that the first 3 digits represent the
-            // country, and that there are nine digits in total.
-            return true;
-        },
-        rules: {
-            multipliers: {},
-            regex: [/^(EU)(\d{9})$/]
-        }
-    };
-});
