--- lib/umd/lib/countries/index.js
@@ -1,85 +1,0 @@
-(function (factory) {
-    if (typeof module === "object" && typeof module.exports === "object") {
-        var v = factory(require, exports);
-        if (v !== undefined) module.exports = v;
-    }
-    else if (typeof define === "function" && define.amd) {
-        define(["require", "exports", "./andorra", "./australia", "./austria", "./belgium", "./brazil", "./bulgaria", "./croatiat", "./cyprus", "./czechRepublic", "./denmark", "./estonia", "./europe", "./finland", "./france", "./germany", "./greece", "./hungary", "./ireland", "./italy", "./latvia", "./lithuania", "./luxembourg", "./malta", "./netherlands", "./norway", "./poland", "./portugal", "./romania", "./russia", "./serbia", "./slovakiaRepublic", "./slovenia", "./spain", "./sweden", "./switzerland", "./unitedKingdom"], factory);
-    }
-})(function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    exports.unitedKingdom = exports.switzerland = exports.sweden = exports.spain = exports.slovenia = exports.slovakiaRepublic = exports.serbia = exports.russia = exports.romania = exports.portugal = exports.poland = exports.norway = exports.netherlands = exports.malta = exports.luxembourg = exports.lithuania = exports.latvia = exports.italy = exports.ireland = exports.hungary = exports.greece = exports.germany = exports.france = exports.finland = exports.europe = exports.estonia = exports.denmark = exports.czechRepublic = exports.cyprus = exports.croatia = exports.bulgaria = exports.brazil = exports.belgium = exports.austria = exports.australia = exports.andorra = void 0;
-    var andorra_1 = require("./andorra");
-    Object.defineProperty(exports, "andorra", { enumerable: true, get: function () { return andorra_1.andorra; } });
-    var australia_1 = require("./australia");
-    Object.defineProperty(exports, "australia", { enumerable: true, get: function () { return australia_1.australia; } });
-    var austria_1 = require("./austria");
-    Object.defineProperty(exports, "austria", { enumerable: true, get: function () { return austria_1.austria; } });
-    var belgium_1 = require("./belgium");
-    Object.defineProperty(exports, "belgium", { enumerable: true, get: function () { return belgium_1.belgium; } });
-    var brazil_1 = require("./brazil");
-    Object.defineProperty(exports, "brazil", { enumerable: true, get: function () { return brazil_1.brazil; } });
-    var bulgaria_1 = require("./bulgaria");
-    Object.defineProperty(exports, "bulgaria", { enumerable: true, get: function () { return bulgaria_1.bulgaria; } });
-    var croatiat_1 = require("./croatiat");
-    Object.defineProperty(exports, "croatia", { enumerable: true, get: function () { return croatiat_1.croatia; } });
-    var cyprus_1 = require("./cyprus");
-    Object.defineProperty(exports, "cyprus", { enumerable: true, get: function () { return cyprus_1.cyprus; } });
-    var czechRepublic_1 = require("./czechRepublic");
-    Object.defineProperty(exports, "czechRepublic", { enumerable: true, get: function () { return czechRepublic_1.czechRepublic; } });
-    var denmark_1 = require("./denmark");
-    Object.defineProperty(exports, "denmark", { enumerable: true, get: function () { return denmark_1.denmark; } });
-    var estonia_1 = require("./estonia");
-    Object.defineProperty(exports, "estonia", { enumerable: true, get: function () { return estonia_1.estonia; } });
-    var europe_1 = require("./europe");
-    Object.defineProperty(exports, "europe", { enumerable: true, get: function () { return europe_1.europe; } });
-    var finland_1 = require("./finland");
-    Object.defineProperty(exports, "finland", { enumerable: true, get: function () { return finland_1.finland; } });
-    var france_1 = require("./france");
-    Object.defineProperty(exports, "france", { enumerable: true, get: function () { return france_1.france; } });
-    var germany_1 = require("./germany");
-    Object.defineProperty(exports, "germany", { enumerable: true, get: function () { return germany_1.germany; } });
-    var greece_1 = require("./greece");
-    Object.defineProperty(exports, "greece", { enumerable: true, get: function () { return greece_1.greece; } });
-    var hungary_1 = require("./hungary");
-    Object.defineProperty(exports, "hungary", { enumerable: true, get: function () { return hungary_1.hungary; } });
-    var ireland_1 = require("./ireland");
-    Object.defineProperty(exports, "ireland", { enumerable: true, get: function () { return ireland_1.ireland; } });
-    var italy_1 = require("./italy");
-    Object.defineProperty(exports, "italy", { enumerable: true, get: function () { return italy_1.italy; } });
-    var latvia_1 = require("./latvia");
-    Object.defineProperty(exports, "latvia", { enumerable: true, get: function () { return latvia_1.latvia; } });
-    var lithuania_1 = require("./lithuania");
-    Object.defineProperty(exports, "lithuania", { enumerable: true, get: function () { return lithuania_1.lithuania; } });
-    var luxembourg_1 = require("./luxembourg");
-    Object.defineProperty(exports, "luxembourg", { enumerable: true, get: function () { return luxembourg_1.luxembourg; } });
-    var malta_1 = require("./malta");
-    Object.defineProperty(exports, "malta", { enumerable: true, get: function () { return malta_1.malta; } });
-    var netherlands_1 = require("./netherlands");
-    Object.defineProperty(exports, "netherlands", { enumerable: true, get: function () { return netherlands_1.netherlands; } });
-    var norway_1 = require("./norway");
-    Object.defineProperty(exports, "norway", { enumerable: true, get: function () { return norway_1.norway; } });
-    var poland_1 = require("./poland");
-    Object.defineProperty(exports, "poland", { enumerable: true, get: function () { return poland_1.poland; } });
-    var portugal_1 = require("./portugal");
-    Object.defineProperty(exports, "portugal", { enumerable: true, get: function () { return portugal_1.portugal; } });
-    var romania_1 = require("./romania");
-    Object.defineProperty(exports, "romania", { enumerable: true, get: function () { return romania_1.romania; } });
-    var russia_1 = require("./russia");
-    Object.defineProperty(exports, "russia", { enumerable: true, get: function () { return russia_1.russia; } });
-    var serbia_1 = require("./serbia");
-    Object.defineProperty(exports, "serbia", { enumerable: true, get: function () { return serbia_1.serbia; } });
-    var slovakiaRepublic_1 = require("./slovakiaRepublic");
-    Object.defineProperty(exports, "slovakiaRepublic", { enumerable: true, get: function () { return slovakiaRepublic_1.slovakiaRepublic; } });
-    var slovenia_1 = require("./slovenia");
-    Object.defineProperty(exports, "slovenia", { enumerable: true, get: function () { return slovenia_1.slovenia; } });
-    var spain_1 = require("./spain");
-    Object.defineProperty(exports, "spain", { enumerable: true, get: function () { return spain_1.spain; } });
-    var sweden_1 = require("./sweden");
-    Object.defineProperty(exports, "sweden", { enumerable: true, get: function () { return sweden_1.sweden; } });
-    var switzerland_1 = require("./switzerland");
-    Object.defineProperty(exports, "switzerland", { enumerable: true, get: function () { return switzerland_1.switzerland; } });
-    var unitedKingdom_1 = require("./unitedKingdom");
-    Object.defineProperty(exports, "unitedKingdom", { enumerable: true, get: function () { return unitedKingdom_1.unitedKingdom; } });
-});
