--- lib/umd/lib/countries/slovakiaRepublic.js
@@ -1,26 +1,0 @@
-(function (factory) {
-    if (typeof module === "object" && typeof module.exports === "object") {
-        var v = factory(require, exports);
-        if (v !== undefined) module.exports = v;
-    }
-    else if (typeof define === "function" && define.amd) {
-        define(["require", "exports"], factory);
-    }
-})(function (require, exports) {
-    "use strict";
-    Object.defineProperty(exports, "__esModule", { value: true });
-    exports.slovakiaRepublic = void 0;
-    exports.slovakiaRepublic = {
-        name: 'Slovakia Republic',
-        codes: ['SK', 'SVK', '703'],
-        calcFn: (vat) => {
-            const expect = 0;
-            const checkDigit = Number(vat) % 11;
-            return checkDigit === expect;
-        },
-        rules: {
-            multipliers: {},
-            regex: [/^(SK)([1-9]\d[2346-9]\d{7})$/]
-        }
-    };
-});
