+++ package.json
@@ -11,7 +11,7 @@
     "build:commonjs": "yarn ts --outDir lib/commonjs --module CommonJS --target ES5",
     "build:es6": "yarn ts --outDir lib/es6 --module ES6",
     "build:amd": "yarn ts --outDir lib/amd --module AMD",
-    "build:umd": "yarn ts --outDir lib/umd --module UMD",
+    "build:umd": "browserify lib/commonjs/index.js --standalone jsvat -o lib/umd/index.js",
     "build:system": "yarn ts --outDir lib/system --module System",
     "prettier": "prettier --write 'src/**/*.ts'",
     "prettier:check": "prettier --check 'src/**/*.ts'",
@@ -61,6 +61,7 @@
     "@babel/preset-env": "7.13.15",
     "babel": "6.23.0",
     "babel-plugin-transform-es2015-modules-commonjs": "6.26.2",
+    "browserify": "17.0.0",
     "husky": "4.3.8",
     "jest": "26.6.3",
     "prettier": "2.2.1",
